Fix property check when not callable
Previously, the property check included a check if the attribute
is callable. Typical properties, being descriptors, are not callable.
A gap in the test coverage allowed this bug to go unnoticed. This
patch removes the callable check in favour of only checking if the
attibute is some subclass of property.